requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.9.24
cloudinary==1.30.0
Django==3.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
importlib-metadata==5.1.0
Markdown==3.3.6
Pillow==9.3.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
zipp==3.11.0
